idk pages idk how this shi works

+++ src/drivers/features/feature.h
@@ -3,13 +3,13 @@
 
 #include<utils.h>
 
-void FeatureCR0Enable(uint64_t FeatureIdx);
-void FeatureCR4Enable(uint64_t FeatureIdx);
-void SetCR3(uint64_t Value);
-uint64_t GetCR3();
+void FeatureCR0Enable(uint32_t FeatureIdx);
+void FeatureCR4Enable(uint32_t FeatureIdx);
+void SetCR3(uint32_t Value);
+uint32_t GetCR3();
 
-bool FeatureIsPresentEDX(uint64_t FeatureIdx);
-bool FeatureIsPresentECX(uint64_t FeatureIdx);
+bool FeatureIsPresentEDX(uint32_t FeatureIdx);
+bool FeatureIsPresentECX(uint32_t FeatureIdx);
 bool *FindFeatures(size_t *Count);
 char *FeatureName(size_t Index);
 void MSRGet(uint32_t msr, uint32_t *lo, uint32_t *hi);
